fix: total experience as int in job reports
add filter in interview results report
add columns in in terviews report

+++ npro/npro/report/rejected_candidates/rejected_candidates.py
@@ -69,7 +69,7 @@ def get_columns(filters):
         {
             "label": _("Total Experience"),
             "fieldname": "applicant_total_experience_cf",
-            "fieldtype": "Float",
+            "fieldtype": "Int",
             "width": 120,
         },
         {
